feat(landing):  switch root
