+++ src/app/page.tsx
@@ -1,57 +1,46 @@
-import AuthButton from '@/components/AuthButton'
-import ConnectSupabaseSteps from '@/components/ConnectSupabaseSteps'
-import SignUpUserSteps from '@/components/SignUpUserSteps'
-import Header from '@/components/Header'
-import { cookies } from 'next/headers'
+import Navbar from '@/components/Navbar'
+import Footer from '@/components/Footer'
+import ProjectGrid from '@/components/ProjectGrid'
 import { createServerClient } from '@/utils/supabase'
-import ThemeToggle from '@/components/ThemeToggle'
+import { cookies } from 'next/headers'
 
 export default async function Index() {
   const cookieStore = cookies()
+  const supabase = createServerClient(cookieStore)
 
-  const canInitSupabaseClient = () => {
-    // This function is just for the interactive tutorial.
-    // Feel free to remove it once you have Supabase connected.
-    try {
-      createServerClient(cookieStore)
-      return true
-    } catch (e) {
-      return false
-    }
+  const { data: projects, error } = await supabase.from('projects').select('*')
+
+  if (error) {
+    console.error('Error al cargar los proyectos:', error.message)
   }
 
-  const isSupabaseConnected = canInitSupabaseClient()
-
   return (
-    <div className="flex w-full flex-1 flex-col items-center gap-20">
-      <nav className="flex h-16 w-full justify-center border-b border-b-foreground/10">
-        <div className="flex w-full max-w-4xl items-center justify-between p-3 text-sm">
-          {isSupabaseConnected && <AuthButton />}
-        </div>
-      </nav>
+    <main className="flex w-full flex-col bg-background text-foreground">
+      {/* Navbar */}
+      <Navbar />
 
-      <div className="flex max-w-4xl flex-1 flex-col gap-20 px-3">
-        <Header />
-        <main className="flex flex-1 flex-col gap-6">
-          <h2 className="mb-4 text-4xl font-bold">Next steps</h2>
-          {isSupabaseConnected ? <SignUpUserSteps /> : <ConnectSupabaseSteps />}
-        </main>
+      {/* Contenido Principal */}
+      <div className="w-full flex-grow">
+        <section className="mx-auto w-full max-w-7xl px-6 py-12">
+          <h1 className="mb-8 text-center font-jetbrains-mono text-3xl">
+            Proyectos Destacados
+          </h1>
+
+          {projects ? (
+            <ProjectGrid projects={projects} />
+          ) : (
+            <p className="text-center text-muted-foreground">
+              No se pudieron cargar los proyectos. Intenta más tarde.
+            </p>
+          )}
+        </section>
       </div>
 
-      <footer className="w-full justify-center border-t border-t-foreground/10 p-8 text-center text-xs">
-        <p className="mb-6">
-          Powered by{' '}
-          <a
-            href="https://supabase.com/?utm_source=create-next-app&utm_medium=template&utm_term=nextjs"
-            target="_blank"
-            className="font-bold hover:underline"
-            rel="noreferrer"
-          >
-            Supabase
-          </a>
-        </p>
-        <ThemeToggle />
-      </footer>
-    </div>
+      {/* Espaciado Adicional */}
+      <div className="h-20"></div>
+
+      {/* Footer */}
+      <Footer />
+    </main>
   )
 }
