+++ src/components/Footer.tsx
@@ -3,8 +3,8 @@ import { FaGithub } from 'react-icons/fa'
 
 const Footer: React.FC = () => {
   return (
-    <footer className="fixed bottom-0 left-0 z-50 w-full bg-background py-6 text-foreground">
-      <div className="container mx-auto flex flex-col items-center justify-between md:flex-row">
+    <footer className="bottom-0 left-0 z-50 w-full bg-background px-6 py-6 text-foreground sm:px-10 lg:px-16">
+      <div className="mx-auto flex w-full flex-col items-center justify-between md:flex-row">
         {/* Logo alineado a la izquierda */}
         <div className="mb-4 md:mb-0">
           <a href="/" target="_blank" rel="noopener noreferrer">
@@ -18,8 +18,9 @@ const Footer: React.FC = () => {
           </a>
         </div>
 
+        {/* Texto alineado a la derecha */}
         <div className="text-center md:text-right">
-          <p className="text-zinc-300 ">made with 💛 by 🍌</p>
+          <p className="text-zinc-300">made with 💛 by 🍌</p>
           <p className="flex items-center justify-center text-sm text-zinc-400 md:justify-end">
             50% human 50% LLM
             <a
