+++ tailwind.config.js
@@ -4,11 +4,6 @@ module.exports = {
   content: ['./src/**/*.{ts,tsx}'],
   prefix: '',
   theme: {
-    container: {
-      center: true,
-      padding: '2rem',
-      screens: { '2xl': '1400px' },
-    },
     extend: {
       fontFamily: {
         sans: ['var(--font-jetbrains-mono)', 'monospace'], // Por defecto en todo
